+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.38.2-0 / Unreleased
   * [Fixed] #1097 Removed incorrect set-function for jQuery cookie
+  * [Fixed] #1108 minExpandLevel: 2 together with table extension does not work
 
 # 2.38.1 / 2022-01-14
   * Replace deprecated jQuery functions: `$.isArray()`, `$.isFunction()`,